Morten arbejder

+++ LNE_Security/LNE_Security/Screens/ProductScreen.cs
@@ -21,6 +21,12 @@ namespace LNE_Security
         {
             this.company = Company;
         }
+
+        public ProductScreen(Company Company, Product Product)
+        {
+            this.company = Company;
+            this.product = Product;
+        }
         protected override void Draw()
         {
             Title = "Product";
@@ -35,48 +41,52 @@ namespace LNE_Security
                 foreach (Product product in products)
                     productListPage.Add(product);
 
-                productListPage.AddColumn("Product Number", "ProductNumber", 10);
+                productListPage.AddColumn("Product Number", "ProductNumber", "Product Number".Length);
                 productListPage.AddColumn("Product Name", "ProductName", 25);
-                productListPage.AddColumn("Amount In Storage", "AmountInStorage");
-                productListPage.AddColumn("Cost Price", "CostPrice");
-                productListPage.AddColumn("Sales Price", "SalesPrice");
+                productListPage.AddColumn("Amount In Storage", "AmountInStorage", "Amount In Storage".Length);
+                productListPage.AddColumn("Cost Price " + company.Currency.ToString(), "CostPrice", "Cost Price ".Length + 3); // TODO: FIX ME!!!
+                productListPage.AddColumn("Sales Price " + company.Currency.ToString(), "SalesPrice", "Sales Price ".Length + 3);
                 if (products.Count == 0)
                     productListPage.Draw();
                 else
                     selectedProduct = productListPage.Select();
+                
+                
+            }
+            if(selectedProduct != null)
+            {
                 Console.WriteLine("Selection : " + selectedProduct.ProductName);
                 Console.WriteLine("Enter - product details");
-            }
-     
-            Console.WriteLine("F1 - New product");
-            Console.WriteLine("F2 - Edit product");
-            Console.WriteLine("F8 - Delete product");
-            Console.WriteLine("F10 - Back");
-            Console.WriteLine("Esc - Close app");
-            switch (Console.ReadKey().Key)
-            {
-                case ConsoleKey.Enter:
-                    ScreenHandler.Display(new ProductDetailsScreen(selectedProduct)); // kunne man kalde ProductDetailsScreen direkte i Display()? ScreenHandler.Display(new ProdcuctDetailsScreen(product)); Erklæring er måske ikke nødvendig.
-                    break;
-                case ConsoleKey.F1:
-                    Database.Instance.NewProduct();
-                    break;
-                case ConsoleKey.F2:
-                    ScreenHandler.Display(new EditProductScreen(selectedProduct));
-                    break;
-                case ConsoleKey.Escape:
-                    Environment.Exit(0);
-                    break;
-                case ConsoleKey.F8:
-                    Database.Instance.DeleteProduct(selectedProduct.PID);
-                    break;
-                case ConsoleKey.F10:
-                    ScreenHandler.Display(new MainMenuScreen(this.company));
-                    break;
-                default:
-                    break;
-            }
-            
+
+                Console.WriteLine("F1 - New product");
+                Console.WriteLine("F2 - Edit product");
+                Console.WriteLine("F8 - Delete product");
+                Console.WriteLine("F10 - Back");
+                Console.WriteLine("Esc - Close app");
+                switch (Console.ReadKey().Key)
+                {
+                    case ConsoleKey.Enter:
+                        ScreenHandler.Display(new ProductDetailsScreen(selectedProduct, company));
+                        break;
+                    case ConsoleKey.F1:
+                        Database.Instance.NewProduct();
+                        break;
+                    case ConsoleKey.F2:
+                        ScreenHandler.Display(new EditProductScreen(selectedProduct));
+                        break;
+                    case ConsoleKey.Escape:
+                        Environment.Exit(0);
+                        break;
+                    case ConsoleKey.F8:
+                        Database.Instance.DeleteProduct(selectedProduct.PID);
+                        break;
+                    case ConsoleKey.F10:
+                        ScreenHandler.Display(new MainMenuScreen(this.company));
+                        break;
+                    default:
+                        break;
+                }
+            }  
         }
     }
 }
